requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1